+++ src/support/kernels/torch_kernel.py
@@ -36,20 +36,21 @@ class TorchKernel(AbstractKernel):
 
     def convolve(self, x, y, p, mode='gaussian'):
         # move tensors to device if needed
+        assert x.device == y.device == p.device, 'x, y and p must be on the same device'
 
         res = None
 
         if mode in ['gaussian', 'pointcloud']:
-            previous_device = x.device.type
-            (x, y, p) = map(self.__move_tensor_to_device_if_needed, [x, y, p])
+            # previous_device = x.device.type
+            # (x, y, p) = map(self.__move_tensor_to_device_if_needed, [x, y, p])
 
             sq = self._squared_distances(x, y)
 
             # if x.type() == 'torch.cuda.FloatTensor'
             res = torch.mm(torch.exp(-sq / (self.kernel_width ** 2)), p)
             # res = torch.mm(1.0 / (1 + sq / self.kernel_width ** 2), p)
-            if previous_device == 'cpu':
-                res = res.cpu()
+            # if previous_device == 'cpu':
+            #     res = res.cpu()
 
         elif mode == 'varifold':
             assert isinstance(x, tuple), 'x must be a tuple'
@@ -57,13 +58,13 @@ class TorchKernel(AbstractKernel):
             assert isinstance(y, tuple), 'y must be a tuple'
             assert len(y) == 2, 'tuple length must be 2'
 
-            previous_device = x[0].device.type
-            (x, y, p) = map(self.__move_tensor_to_device_if_needed, [x, y, p])
+            # previous_device = x[0].device.type
+            # (x, y, p) = map(self.__move_tensor_to_device_if_needed, [x, y, p])
 
             sq = self._squared_distances(x[0], y[0])
             res = torch.mm(gaussian(sq, self.kernel_width) * binet(torch.mm(x[1], torch.t(y[1]))), p)
-            if previous_device == 'cpu':
-                res = res.cpu()
+            # if previous_device == 'cpu':
+            #     res = res.cpu()
         else:
             raise RuntimeError('Unknown kernel mode.')
 
@@ -118,7 +119,7 @@ class TorchKernel(AbstractKernel):
 
         def move(t, device):
             res = t.to(device=device)
-            assert res.device.type == self.device, 'error moving tensor to device'
+            # assert res.device.type == self.device, 'error moving tensor to device'
             return res
 
         res = None